do not use uintptr as hash key

+++ src/platforms/platform_impl2_emscripten_sdl2_renderer_webgl1.h
@@ -31,9 +31,9 @@ struct Platform2_Emscripten_SDL2_Renderer_WebGL1
     bool webgl_context_ready;
     struct Pix3DGL* pix3dgl;
     int next_model_index;
-    std::unordered_map<uintptr_t, int> model_index_by_key;
-    std::unordered_set<uintptr_t> loaded_model_keys;
-    std::unordered_set<uintptr_t> loaded_scene_element_keys;
+    std::unordered_map<uint64_t, int> model_index_by_key;
+    std::unordered_set<uint64_t> loaded_model_keys;
+    std::unordered_set<uint64_t> loaded_scene_element_keys;
     std::unordered_set<int> loaded_texture_ids;
 
     std::unordered_map<struct DashPixFont*, GLFontAtlasEntryES2> font_atlas_cache;
