+++ src/platforms/platform_impl2_osx_sdl2_renderer_opengl3.h
@@ -31,9 +31,9 @@ struct Platform2_OSX_SDL2_Renderer_OpenGL3
     bool gl_context_ready;
     struct Pix3DGL* pix3dgl;
     int next_model_index;
-    std::unordered_map<uintptr_t, int> model_index_by_key;
-    std::unordered_set<uintptr_t> loaded_model_keys;
-    std::unordered_set<uintptr_t> loaded_scene_element_keys;
+    std::unordered_map<uint64_t, int> model_index_by_key;
+    std::unordered_set<uint64_t> loaded_model_keys;
+    std::unordered_set<uint64_t> loaded_scene_element_keys;
     std::unordered_set<int> loaded_texture_ids;
 
     std::unordered_map<struct DashPixFont*, GLFontAtlasEntry> font_atlas_cache;
